Fixed errors. Removed the UserData class and created Register Model instead. Added Javax validation.
