+++ pom.xml
@@ -40,6 +40,10 @@
 			<scope>test</scope>
 		</dependency>
 	
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-validation</artifactId>
+		</dependency>
 	</dependencies>
 
 	<build>
