+++ src/main/resources/templates/register.html
@@ -8,7 +8,7 @@
   <body>
 
 	<h2 th:text="${message}">Welcome to Preach Blogs </h2><br/>
-	<form action="#" th:action="@{doLogin}" th:object="${loginModel}" method="post">
+	<form action="#" th:action="@{doRegister}" th:object="${registerModel}" method="post">
 		<table>
 			<tr>
 				<td>First Name:</td><td><input type="text" th:field="*{userFirstName}" ></td>
